requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
psycopg2-binary
geoalchemy2
shapely
pydantic
python-dotenv
celery[redis]
redis
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dateutil
requests
beautifulsoup4
tweepy
nltk
textblob
pillow
aiofiles
scikit-learn
